+++ sdk/index.d.ts
@@ -4,7 +4,6 @@
 
 export type ScrapeMode =
     | 'headings-paragraphs'
-    | 'genesis-original'
     | 'all-text'
     | 'articles'
     | 'lists'